recipe repositories and recipe creation screen

+++ app/src/main/java/week/on/a/plate/data/dataView/recipe/IngredientInRecipeView.kt
@@ -7,5 +7,5 @@ data class IngredientInRecipeView(
     val id: Long,
     val ingredientView: IngredientView,
     val description:String,
-    val count:Double
+    val count:Int
 )
