+++ app/src/main/java/week/on/a/plate/screenRecipeDetails/event/RecipeDetailsEvent.kt
@@ -11,5 +11,5 @@ sealed class RecipeDetailsEvent: Event() {
     data object AddToMenu:RecipeDetailsEvent()
     data object PlusPortionsView:RecipeDetailsEvent()
     data object MinusPortionsView:RecipeDetailsEvent()
-    data class StartTimerForStep(val time: Long, val act: Context) : RecipeDetailsEvent()
+    data class StartTimerForStep(val time: Int, val act: Context) : RecipeDetailsEvent()
 }
